Split Matcher into a separate file
I suspect that tracing the various paths through this, or spotting when
we're breaking encapsulation, will be easier if this is at one level
removed.
